Tests/LibWeb/Text/expected/css/filter-hue-rotate-unitless-zero.txt

hue-rotate(0.0): hue-rotate(0deg)
hue-rotate(0e0): hue-rotate(0deg)
